Mongodb test (#59)
* template test

* Restructure mongodb
1.Amend SetCpuCoreLimit function to support set cpu num with fraction.
2.Support default partition,which would be shown with a '*' at the end.
3.Add account manager module in file. AccountManager.cpp and AccountManager.h.
4.Replace the flow constructor with the basic constructor.
5.Restructure DbClient by DAO fabric.
6.Support for adding, deleting, searching and modifying accounts and users

* Mongodb transaction and mutex
1.Add boost read and write mutex example to AddUser().
2.User transaction to operate document atomically.
3.Configure mongodb replica sets.
4.Support for adding, deleting, searching and modifying qos.
5.Simplify part of the code.
6.Deal with some warning.
7.Catch mongodb exception when connecting.
8.Add some common api in DbClient.

* read and write mutex
1.Using rw_mutex to avoid access conflicts to maps.
2.Add transaction to recurrence modify operation.
3.Fix bugs in account manager.
4.Rename some private function.
5.Reconstruct part of APIs in account manager,some unlocked APIs have been added for external calls.
6.Integrate the modifies to the map into a module and place it after the database modifies.
7.Delete redundant 'Ctld::' statement.
8.Use mongocxx::pool to ensure thread safety.
9.Catch exception when convert view to struct

* thread_local variable

* Accept some suggestions
1.Support customization of replica set name in config file.
2.Rename some variables.
3.Normalize parameter transfer.
4.Add thread local variables

* Add blank line to divide code && Rename some variables and functions

* Rename private function in class && Avoid null Node in yaml file to add to g_config

* Use thread local variable replace the session as a parameter && delete the member variable in class MongodbClient to ensure thread safety.

* Continue to amend

* Continue to amend
1.Use lock_guard replace lock and unlock.
2.GetXXXInfoNoLock_ No longer copy value, but transfer the pointer externally.
3.GetXXXInfo external transfer of pointers containing mutex and data.
4.Capitalize variable members in MongodbClient::EntityType.
5.Use std::move to avoid redundant replication operations.
6.Standardized the reference of enum variable in grpc.
7.Limits the scope of kvp.
8.Removed redundant namespace declarations 'Ctld::'.
7.Extract the common enum definition in proto.

* Rename ScopeExclusiveSharedPtr to ScopeSharedPtr && Adjusted the logic of creation_time and mod_time

+++ src/Utilities/PublicHeader/include/crane/PublicHeader.h
@@ -173,7 +173,7 @@ struct fmt::formatter<CranedId> {
 // Model the allocatable resources on a craned node.
 // It contains CPU and memory by now.
 struct AllocatableResource {
-  uint32_t cpu_count = 0;
+  double cpu_count = 0.0;
 
   // See documentation of cgroup memory.
   // https://www.kernel.org/doc/Documentation/cgroup-v1/memory.txt
